Refactoring
(Code not working)
Refactoring my first game in Unity. Trying to separate the monolithic class Player. Setting up namespaces to see dependencies and decoupling them.
To Do:
Player, PlayerDamage, Asteroid depends on Managers.
PowerUp and depends on PlayerModules.
Player needs to know about the state of the powerups (booleans).
PlayerDamage needs to know about the state of the shield powerup and the shieldvisuals from the powerups(boolean and GameObject).
Extract ShipMovement and Boundaries from Player into PlayerMovement? (if true) => PlayerMovement will need to know about the state of powerups(boolean) and then we can rename Player to Shooting.
Remove AddScore and place it in a separate class?(TBD)